+++ presidio-analyzer/presidio_analyzer/input_validation/yaml_recognizer_models.py
@@ -151,9 +151,8 @@ class PredefinedRecognizerConfig(BaseRecognizerConfig):
     aggregation_strategy: Optional[str] = Field(
         None, description="Aggregation strategy"
     )
-    chunk_overlap_size: Optional[int] = Field(None, description="Chunk overlap size")
+    chunk_overlap: Optional[int] = Field(None, description="Chunk overlap")
     chunk_size: Optional[int] = Field(None, description="Chunk size")
-    batch_size: Optional[int] = Field(None, description="Batch size")
     max_text_length: Optional[int] = Field(None, description="Max text length")
     device: Optional[Union[str, int]] = Field(None, description="Device (cpu/gpu)")